Gold and GoldDrop classes, attached to enemy prefabs
Gold and GoldDrop classes, attached to enemy prefabs. New gold prefab. Enemies drop gold, hero can pick it up on contact.
