+++ stc_game/Assets/Prefabs/Characters/Character_Goblin_Male_Complete.prefab
@@ -545,6 +545,7 @@ GameObject:
   - component: {fileID: 721315338923108770}
   - component: {fileID: 9118579321821184893}
   - component: {fileID: 230125693817955390}
+  - component: {fileID: 5690782258585447086}
   m_Layer: 0
   m_Name: Character_Goblin_Male_Complete
   m_TagString: Enemy
@@ -670,11 +671,14 @@ MonoBehaviour:
   startState: {fileID: 11400000, guid: bab1dd25e9db19c409b0df1f47b5f795, type: 2}
   currentState: {fileID: 0}
   remainState: {fileID: 11400000, guid: 08d96f2ec4aa4b44a885d459d6763042, type: 2}
+  aggroState: {fileID: 0}
   aggroSergent: {fileID: 0}
   target: {fileID: 0}
   cam: {fileID: 0}
   destination: {fileID: 0}
   waypointList: []
+  playerSaveData: {fileID: 0}
+  startingPositionName: 
   characterStats: {fileID: 0}
   animator: {fileID: 0}
   stateTimeElapsed: 0
@@ -696,6 +700,9 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   presetStats: {fileID: 0}
+  level: 0
+  XP: 0
+  gold: 0
   maxHP: 100
   maxAP: 0
   currentHP: 100
@@ -724,6 +731,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   touching: []
+--- !u!114 &5690782258585447086
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 114438}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 447eef31cd5938448ba1931cda6ed0a8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gold: {fileID: 138872, guid: 4fd0c7547143bde4c851c20099915676, type: 3}
+  offset: {x: 0.5, y: 0, z: 0.5}
 --- !u!1 &115104
 GameObject:
   m_ObjectHideFlags: 0
