+++ stc_game/Assets/Prefabs/Characters/Character_Goblin_WarChief_Complete.prefab
@@ -3359,6 +3359,7 @@ GameObject:
   - component: {fileID: 5492319422314749400}
   - component: {fileID: 6529141008692494347}
   - component: {fileID: 7318591488769944929}
+  - component: {fileID: 1431980808679365883}
   m_Layer: 0
   m_Name: Character_Goblin_WarChief_Complete
   m_TagString: Enemy
@@ -3499,11 +3500,14 @@ MonoBehaviour:
   startState: {fileID: 11400000, guid: bab1dd25e9db19c409b0df1f47b5f795, type: 2}
   currentState: {fileID: 0}
   remainState: {fileID: 11400000, guid: 08d96f2ec4aa4b44a885d459d6763042, type: 2}
+  aggroState: {fileID: 0}
   aggroSergent: {fileID: 0}
   target: {fileID: 0}
   cam: {fileID: 0}
   destination: {fileID: 0}
   waypointList: []
+  playerSaveData: {fileID: 0}
+  startingPositionName: 
   characterStats: {fileID: 0}
   animator: {fileID: 0}
   stateTimeElapsed: 0
@@ -3525,6 +3529,9 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   presetStats: {fileID: 0}
+  level: 0
+  XP: 0
+  gold: 0
   maxHP: 100
   maxAP: 0
   currentHP: 100
@@ -3553,6 +3560,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   touching: []
+--- !u!114 &1431980808679365883
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 173802}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 447eef31cd5938448ba1931cda6ed0a8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gold: {fileID: 138872, guid: 4fd0c7547143bde4c851c20099915676, type: 3}
+  offset: {x: 0.5, y: 0, z: 0.5}
 --- !u!1 &174376
 GameObject:
   m_ObjectHideFlags: 0
