+++ stc_game/Assets/Prefabs/Gold1.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4fd0c7547143bde4c851c20099915676
+timeCreated: 1508373733
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
