+++ stc_game/Assets/SM_Item_Coins_05.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76cc2a6c57d38764fbe93e2749416a94
+timeCreated: 1508373733
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
