+++ stc_game/Assets/Scripts/CharacterBehavior/ActionScripts/DeathAction.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using Assets.Scripts.CharacterBehavior.Drops;
 using UnityEngine;
 
 [CreateAssetMenu(menuName = "PluggableAI/Actions/Death")]
@@ -12,10 +11,13 @@ public class DeathAction : CharacterAction
 
     public void Die(StateController controller)
     {
+        if (controller.GetComponent<GoldDrop>())
+            controller.GetComponent<GoldDrop>().DropGold();
         controller.animator.SetBool("Dead", true);
         controller.gameObject.layer = 2;
         controller.navMeshAgent.enabled = false;
         controller.GetComponent<Rigidbody>().isKinematic = true;
         controller.GetComponent<CapsuleCollider>().enabled = false;
+
     }
 }
