--- stc_game/Assets/Scripts/CharacterBehavior/BaseClasses/BehaviorStats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 906d839498b35d14691445d6822b7d3e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
