+++ stc_game/Assets/Scripts/CharacterBehavior/Drops.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6569a008e9118b843b7af4768200bbc9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
