+++ stc_game/Assets/Scripts/CharacterBehavior/Drops/DropBag.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using UnityEngine;
+
+namespace Assets.Scripts.CharacterBehavior.Drops
+
+{
+    public class DropBag : MonoBehaviour
+    {
+        public GameObject[] items;
+        
+    }
+}
