+++ stc_game/Assets/Scripts/CharacterBehavior/Drops/DropBag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9995147150fe81142837d248d4d31487
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
