+++ stc_game/Assets/Scripts/CharacterBehavior/Drops/Gold.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3549b3e1246ea674680fd8acf41f5f2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
