+++ stc_game/Assets/Scripts/CharacterBehavior/Drops/GoldDrop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 447eef31cd5938448ba1931cda6ed0a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
