+++ stc_game/Assets/Scripts/Combat/ScriptableObjects/StatsPresets/EnemyBaseStatsPreset.asset
@@ -12,6 +12,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ddd6f3dbfa5fa0b40afcfdc411e46da6, type: 3}
   m_Name: EnemyBaseStatsPreset
   m_EditorClassIdentifier: 
+  level: 1
+  gold: 10
+  XP: 5
   maxHP: 100
   maxAP: 100
   strength: 10
