+++ stc_game/Assets/Scripts/Stats.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db297fd87108ec54c8529289af9aae8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
