+++ stc_game/Assets/Scripts/Stats/BehaviorStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 906d839498b35d14691445d6822b7d3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
