+++ stc_game/Assets/Scripts/Stats/CharacterStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f2713685d9b90c4897e1f19bea78896
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
