+++ stc_game/Assets/Scripts/Stats/StatsPreset.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddd6f3dbfa5fa0b40afcfdc411e46da6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
